Modified : file

+++ moveFile.sh
@@ -3,7 +3,10 @@ for file in `ls *.txt`
 do
 	folderName=`echo $file | awk -F . '{print $1}'`;
 	echo $folderName;
-   	rm -r $folderName;
+	if [ -d $folderName ]
+	then
+   		rm -r $folderName;
+	fi
 	mkdir $folderName;
    cp $file $folderName;
    echo Copied $file into $folderName
